components: major reorganization
