+++ src/components/pages/StakePage.css
