+++ src/components/pages/StakePage.js
